added labels to signup form input fields to adhere to semantic HTML guidelines
